XW-1779: traits
