+++ includes/wikia/models/DesignSystem/LineTextObject.class.php
@@ -1,17 +1,9 @@
 <?php
 
 class LineTextObject {
+	use WdsTitleTrait;
+
 	const TYPE = 'line-text';
-	private $title;
-
-	public function __construct() {
-	}
-
-	public function setTranslatableTitle( $key ) {
-		$this->title = ( new TranslatableTextObject( $key ) )->get();
-
-		return $this;
-	}
 
 	public function get() {
 		return [
