+++ includes/wikia/models/DesignSystem/LinkBrandedObject.class.php
@@ -1,35 +1,15 @@
 <?php
 
 class LinkBrandedObject {
-	const TYPE = 'link-branded';
-	private $brand;
-	private $title;
-	private $href;
-	private $trackingLabel;
+	use WdsTitleTrait;
+	use WdsLinkTrait;
 
-	public function __construct() {
-	}
+	const TYPE = 'link-branded';
+
+	private $brand;
 
 	public function setBrand( $brand ) {
 		$this->brand = $brand;
-
-		return $this;
-	}
-
-	public function setTranslatableTitle( $key ) {
-		$this->title = ( new TranslatableTextObject( $key ) )->get();
-
-		return $this;
-	}
-
-	public function setHref( $href ) {
-		$this->href = $href;
-
-		return $this;
-	}
-
-	public function setTrackingLabel( $label ) {
-		$this->trackingLabel = $label;
 
 		return $this;
 	}
