+++ includes/wikia/models/DesignSystem/WdsTitleTrait.php
@@ -1,0 +1,17 @@
+<?php
+
+trait WdsTitleTrait {
+	private $title;
+
+	public function setTitle( $value ) {
+		$this->title = ( new TextObject( $value ) )->get();
+
+		return $this;
+	}
+
+	public function setTranslatableTitle( $key ) {
+		$this->title = ( new TranslatableTextObject( $key ) )->get();
+
+		return $this;
+	}
+}
